forenkle patterns, gå vekk fra generiske kall osv

+++ src/main/kotlin/no/nav/sokos/os/ekstern/api/dto/vedtak/TilbakekrevingsvedtakResponse.kt
@@ -6,6 +6,6 @@ import kotlinx.serialization.Serializable
 data class TilbakekrevingsvedtakResponse(
     val status: Int,
     val melding: String,
-    val vedtakId: Long,
+    val vedtakId: Int,
     val datoVedtakFagsystem: String,
 )
